requirements.txt
fastapi
uvicorn
python-multipart
tensorflow-cpu
opencv-python-headless
numpy
pillow
pandas
streamlit
requests